Add links and endpoints for partners
New routes and views were added for the partner endpoints. The url parameters are now passed to the response template as well, since the partner_id is needed to filter the correct partner from the response object. Related Detail and List reponse types were added to allow for easy filtering on a specific embedded relation (in this case ‘partners’).